# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Amina API
  version: 1.0.0
  description: |
    Image generation and utilities API for the Amina Discord Bot.

    ## Authentication

    All `/v1/*` endpoints require API key authentication via Bearer token.

    ```bash
    curl -H "Authorization: Bearer amina_YOUR_API_KEY" https://api.4mina.app/v1/images/rank-card?...
    ```

    Get your API key at [api.4mina.app/dashboard](https://api.4mina.app/dashboard)

    ## Rate Limiting

    - **60 requests per minute** per API key
    - Rate limit headers included in responses:
      - `X-RateLimit-Limit`: Maximum requests per window
      - `X-RateLimit-Remaining`: Requests remaining
      - `X-RateLimit-Reset`: Unix timestamp when limit resets

    ## Response Format

    All JSON responses follow this structure:

    ```json
    {
      "success": true,
      "data": { ... },
      "meta": {
        "generatedAt": "2024-01-01T00:00:00.000Z"
      }
    }
    ```

    Image endpoints return SVG directly with `Content-Type: image/svg+xml`.

  contact:
    name: Amina Support
    url: https://discord.gg/4mina

  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.4mina.app
    description: Production
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: API key from dashboard
  parameters:
    imageUrl:
      name: image
      in: query
      required: true
      schema:
        type: string
        format: uri
      description: URL of the image to process
    width:
      name: width
      in: query
      schema:
        type: integer
        default: 512
        minimum: 1
        maximum: 1024
      description: Output width in pixels
    height:
      name: height
      in: query
      schema:
        type: integer
        default: 512
        minimum: 1
        maximum: 1024
      description: Output height in pixels
  responses:
    SvgImage:
      description: SVG image
      content:
        image/svg+xml:
          schema:
            type: string
    BadRequest:
      description: Bad request - missing or invalid parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized - invalid or missing API key
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            message:
              type: string
            code:
              type: string
        meta:
          type: object
          properties:
            generatedAt:
              type: string
              format: date-time
paths:
...
  /v1/images/filters/greyscale:
    get:
      tags: [Filters]
      summary: Convert image to greyscale
      parameters:
        - $ref: '#/components/parameters/imageUrl'
        - $ref: '#/components/parameters/width'
        - $ref: '#/components/parameters/height'
        - name: intensity
          in: query
          schema:
            type: number
            minimum: 0
            maximum: 1
            default: 1
          description: Filter intensity
      responses:
        '200':
          $ref: '#/components/responses/SvgImage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /v1/images/filters/blur:
    get:
      tags: [Filters]
      summary: Apply blur effect
      parameters:
        - $ref: '#/components/parameters/imageUrl'
        - $ref: '#/components/parameters/width'
        - $ref: '#/components/parameters/height'
        - name: radius
          in: query
          schema:
            type: number
            minimum: 0
            maximum: 50
            default: 5
          description: Blur radius
      responses:
        '200':
          $ref: '#/components/responses/SvgImage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /v1/images/filters/sepia:
    get:
      tags: [Filters]
      summary: Apply sepia filter
      parameters:
        - $ref: '#/components/parameters/imageUrl'
        - $ref: '#/components/parameters/width'
        - $ref: '#/components/parameters/height'
        - name: intensity
          in: query
          schema:
            type: number
            minimum: 0
            maximum: 1
            default: 1
      responses:
        '200':
          $ref: '#/components/responses/SvgImage'
  /v1/images/filters/invert:
    get:
      tags: [Filters]
      summary: Invert image colors
      parameters:
        - $ref: '#/components/parameters/imageUrl'
        - $ref: '#/components/parameters/width'
        - $ref: '#/components/parameters/height'
      responses:
        '200':
          $ref: '#/components/responses/SvgImage'
  /v1/images/filters/brighten:
    get:
      tags: [Filters]
      summary: Brighten an image
      parameters:
        - $ref: '#/components/parameters/imageUrl'
        - $ref: '#/components/parameters/width'
        - $ref: '#/components/parameters/height'
        - name: amount
          in: query
          schema:
            type: number
            minimum: 0
            maximum: 2
            default: 1.3
      responses:
        '200':
          $ref: '#/components/responses/SvgImage'
  /v1/images/filters/darken:
    get:
      tags: [Filters]
      summary: Darken an image
      parameters:
        - $ref: '#/components/parameters/imageUrl'
        - $ref: '#/components/parameters/width'
        - $ref: '#/components/parameters/height'
        - name: amount
          in: query
          schema:
            type: number
            minimum: 0
            maximum: 1
            default: 0.7
      responses:
        '200':
          $ref: '#/components/responses/SvgImage'

  # ========== Overlays ==========
...